test_requirements.txt
django-nose
coverage
django-libs
django-coverage
ipdb
flake8
fabric
factory_boy
mock
south
